requirements.txt
Django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
django-mysql==4.11.0
django-cors-headers==4.2.0
pika==1.3.2